paymentOrder amount 필드 BigDecimal -> Long 수정

+++ payment-service/src/main/java/com/example/backend/application/service/CheckoutService.java
@@ -43,7 +43,7 @@ public class CheckoutService implements CheckoutUseCase {
                                 .sellerId(product.getSellerId())
                                 .orderId(command.idempotencyKey())
                                 .productId(product.getId())
-                                .amount(product.getAmount())
+                                .amount(product.getAmount().longValue())
                                 .paymentStatus(PaymentStatus.NOT_STARTED)
                                 .build())
                         .toList())
