+++ payment-service/src/main/java/com/example/backend/domain/PaymentOrder.java
@@ -13,14 +13,14 @@ public class PaymentOrder {
     private Long buyerId;
     private Long productId;
     private String orderId;
-    private BigDecimal amount;
+    private Long amount;
     private PaymentStatus paymentStatus;
     private boolean isLedgerUpdated = false;
     private boolean isWalletUpdated = false;
     private boolean isPaymentDone = false;
 
     @Builder
-    public PaymentOrder(Long id, Long paymentEventId, Long sellerId, Long buyerId, Long productId, String orderId, BigDecimal amount, PaymentStatus paymentStatus, boolean isLedgerUpdated, boolean isWalletUpdated, boolean isPaymentDone) {
+    public PaymentOrder(Long id, Long paymentEventId, Long sellerId, Long buyerId, Long productId, String orderId, Long amount, PaymentStatus paymentStatus, boolean isLedgerUpdated, boolean isWalletUpdated, boolean isPaymentDone) {
         this.id = id;
         this.paymentEventId = paymentEventId;
         this.sellerId = sellerId;
